+++ packages/react/AppShell/src/appShell.types.ts
@@ -6,7 +6,8 @@ export interface AppShellComponents {
   Header: typeof AppShellHeader;
 }
 
-export interface AppShellProps extends HTMLAttributes<HTMLElement> {
+export interface AppShellProps
+  extends Omit<HTMLAttributes<HTMLElement>, "color"> {
   /** Content for the body of the application */
   children: ReactNode;
   /** Optional slot for left sidebar menu */