+++ packages/react/InteractiveList/src/components/InteractiveListButtonItem/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListButtonItem } from "./InteractiveListButtonItem";
+
+export { InteractiveListButtonItem } from "./InteractiveListButtonItem";
+export default InteractiveListButtonItem;