+++ packages/react/InteractiveList/src/components/InteractiveListRadioItem/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListRadioItem } from "./InteractiveListRadioItem";
+
+export { InteractiveListRadioItem } from "./InteractiveListRadioItem";
+export default InteractiveListRadioItem;