+++ packages/react/InteractiveList/src/index.ts
@@ -1,0 +1,5 @@
+import { InteractiveList } from "./InteractiveList";
+
+export { InteractiveList } from "./InteractiveList";
+export type { InteractiveListProps } from "./interactiveList.types";
+export default InteractiveList;