+++ packages/react/Menu/src/components/MenuBody/menuBody.types.ts
@@ -1,6 +1,7 @@
 import { ReactNode, HTMLAttributes } from "react";
 
-export interface MenuBodyProps extends HTMLAttributes<HTMLElement> {
+export interface MenuBodyProps
+  extends Omit<HTMLAttributes<HTMLElement>, "color"> {
   /** Content of the menu body */
   children: ReactNode;
 }