+++ packages/react/Menu/src/components/MenuSection/menuSection.types.ts
@@ -1,6 +1,7 @@
 import { HTMLAttributes, ReactNode } from "react";
 
-export interface MenuSectionProps extends HTMLAttributes<HTMLElement> {
+export interface MenuSectionProps
+  extends Omit<HTMLAttributes<HTMLElement>, "color"> {
   /** Optional title of the section */
   title?: string;
   /** Content of the menu section */